Document that auto-extensions registered via JNI will fail if they open a db (to avoid triggering an endless loop in the auto-extensions). Discover the hard way that JDKv19 creates different mangled JNI names for some functions than JDKv8. Start reformatting much of the JNI API decls in prep for making them even longer.
FossilOrigin-Name: 99c0941f1c006622932a9cca12661f354f363a6c8a2b5675ea66149e0a9eb927

+++ ext/jni/src/org/sqlite/jni/tester/SQLTester.java
@@ -228,10 +228,13 @@ public class SQLTester {
   }
 
   private void setupInitialDb() throws DbException {
-    outln("setupInitialDb()");
-    closeDb(0);
-    Util.unlink(initialDbName);
-    openDb(0, initialDbName, true);
+    if( null==aDb[0] ){
+      Util.unlink(initialDbName);
+      openDb(0, initialDbName, true);
+    }else{
+      outln("WARNING: setupInitialDb() unexpectedly ",
+            "triggered while it is opened.");
+    }
   }
 
   static final String[] startEmoji = {
@@ -594,7 +597,6 @@ public class SQLTester {
     final AutoExtension ax = new AutoExtension() {
         private final SQLTester tester = t;
         public int xEntryPoint(sqlite3 db){
-          tester.outln("AutoExtension running db init code on ",db);
           final String init = tester.getDbInitSql();
           if( !init.isEmpty() ){
             tester.execSql(db, true, ResultBufferMode.NONE, null, init);
